Complete refactor of InstrumentPage component
Moved helper functions into src/helpers/functions.ts
Updated import statements.
